# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meta Earth Wallet Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all wallets
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
    post:
      operationId: createWallet
      tags: [wallets]
      summary: Add a wallet (mnemonic encrypted and stored)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletInput"
      responses:
        "201":
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "400":
          description: Invalid input
  /wallets/{id}:
    get:
      operationId: getWallet
      tags: [wallets]
      summary: Get a wallet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "404":
          description: Not found
    delete:
      operationId: deleteWallet
      tags: [wallets]
      summary: Remove a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /wallets/{id}/balance:
    get:
      operationId: getWalletBalance
      tags: [wallets]
      summary: Fetch live MEC balance from blockchain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balance info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
        "404":
          description: Not found
  /wallets/{id}/transactions:
    get:
      operationId: getWalletTransactions
      tags: [wallets]
      summary: Fetch on-chain transaction history for a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransactionList"
        "404":
          description: Not found
  /wallets/{id}/staking-rewards:
    get:
      operationId: getWalletStakingRewards
      tags: [wallets]
      summary: Fetch pending (withdrawable) staking / block rewards for a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staking rewards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletStakingRewards"
        "404":
          description: Not found
  /wallets/bulk-import:
    post:
      operationId: bulkImportWallets
      tags: [wallets]
      summary: Import multiple wallets from a block of mnemonic phrases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportInput"
      responses:
        "200":
          description: Import summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
        "400":
          description: Invalid input
...
components:
  schemas:
...
    Wallet:
      type: object
      required: [id, label, address, network, createdAt]
      properties:
        id:
          type: integer
        label:
          type: string
        address:
          type: string
        network:
          type: string
        verified:
          type: boolean
        importSource:
          type: string
        createdAt:
          type: string
          format: date-time
    WalletInput:
      type: object
      required: [label, password, network]
      properties:
        label:
          type: string
          minLength: 1
        mnemonic:
          type: string
          minLength: 1
        privateKey:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        network:
          type: string
          default: mainnet
    WalletBalance:
      type: object
      required: [walletId, address, balance, denom]
      properties:
        walletId:
          type: integer
        address:
          type: string
        balance:
          type: string
        denom:
          type: string
        usdValue:
          type: ["string", "null"]
...
    WalletStakingRewards:
      type: object
      required: [walletId, address, rewards, totalMEC, totalUMec]
      properties:
        walletId:
          type: integer
        address:
          type: string
        rewards:
          type: array
          items:
            type: object
            required: [validatorAddress, amount, amountRaw]
            properties:
              validatorAddress:
                type: string
              amount:
                type: string
              amountRaw:
                type: string
        totalMEC:
          type: string
        totalUMec:
          type: string
    BulkImportInput:
      type: object
      required: [mnemonics, password]
      properties:
        mnemonics:
          type: string
          description: One mnemonic phrase per line (12 or 24 words each)
        password:
          type: string
          minLength: 1
        network:
          type: string
          default: mainnet
    BulkImportResult:
      type: object
      required: [total, verified, unverified, skipped]
      properties:
        total:
          type: integer
        verified:
          type: integer
        unverified:
          type: integer
        skipped:
          type: integer
        wallets:
          type: object
          properties:
            verified:
              type: array
              items:
                type: object
                properties:
                  address:
                    type: string
                  label:
                    type: string
                  walletId:
                    type: integer
            unverified:
              type: array
              items:
                type: object
                properties:
                  address:
                    type: string
                  label:
                    type: string
                  walletId:
                    type: integer
        skippedDetails:
          type: array
          items:
            type: object
            properties:
              reason:
                type: string
              phrase:
                type: string
...
    WalletTransaction:
      type: object
      required: [txHash, height, timestamp, direction, amount, amountRaw, counterpart, memo, success]
      properties:
        txHash:
          type: string
        height:
          type: integer
        timestamp:
          type: string
        direction:
          type: string
          enum: [sent, received]
        amount:
          type: string
        amountRaw:
          type: string
        counterpart:
          type: string
        memo:
          type: string
        success:
          type: boolean
    WalletTransactionList:
      type: object
      required: [walletId, address, transactions]
      properties:
        walletId:
          type: integer
        address:
          type: string
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
...