# lib/api-zod/src/generated/types/bulkImportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meta Earth Wallet Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkImportInput {
  /** One mnemonic phrase per line (12 or 24 words each) */
  mnemonics: string;
  /** @minLength 1 */
  password: string;
  network?: string;
}