# lib/api-zod/src/generated/types/bulkImportResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meta Earth Wallet Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkImportResultSkippedDetailsItem } from './bulkImportResultSkippedDetailsItem';
import type { BulkImportResultWallets } from './bulkImportResultWallets';

export interface BulkImportResult {
  total: number;
  verified: number;
  unverified: number;
  skipped: number;
  wallets?: BulkImportResultWallets;
  skippedDetails?: BulkImportResultSkippedDetailsItem[];
}