# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meta Earth Wallet Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Wallet {
  id: number;
  label: string;
  address: string;
  network: string;
  verified?: boolean;
  importSource?: string;
  createdAt: Date;
}