# lib/api-zod/src/generated/types/walletStakingRewards.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meta Earth Wallet Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletStakingRewardsRewardsItem } from './walletStakingRewardsRewardsItem';

export interface WalletStakingRewards {
  walletId: number;
  address: string;
  rewards: WalletStakingRewardsRewardsItem[];
  totalMEC: string;
  totalUMec: string;
}